refactor to use `getFinancialYearRanges` created externally
